requirements.txt
seaborn==0.11.0
numpy==1.19.2
tensorflow==2.5.0
matplotlib==3.3.2
streamlit==0.81.1
opencv_python==4.5.2.52
pandas==1.1.3
keras_nightly==2.6.0.dev2021052300
keras==2.4.3
Pillow==8.2.0
protobuf==3.17.1
scikit_learn==0.24.2
